refine get color
